Refactor odds helpers into module
