+++ bet_logger.py
@@ -14,20 +14,7 @@ else:  # pragma: no cover - POSIX
     import fcntl
 
 from bankroll import calculate_bet_size
-
-
-def american_odds_to_payout(odds: float) -> float:
-    """Return the profit on a $1 bet for the given American odds."""
-    if odds > 0:
-        return odds / 100.0
-    return 100.0 / abs(odds)
-
-
-def american_odds_to_prob(odds: float) -> float:
-    """Convert American odds to an implied win probability."""
-    if odds > 0:
-        return 100 / (odds + 100)
-    return abs(odds) / (abs(odds) + 100)
+from odds_utils import american_odds_to_prob, american_odds_to_payout
 
 
 def _load_logs(path: Path) -> list[dict]:
