+++ odds_utils.py
